WIP: refatoracao dos atributos de volume para integer

+++ db/migrate/20200212170629_create_carrier_options.rb
@@ -1,8 +1,8 @@
 class CreateCarrierOptions < ActiveRecord::Migration[6.0]
   def change
     create_table :carrier_options do |t|
-      t.float :min_vol
-      t.float :max_vol
+      t.integer :min_vol
+      t.integer :max_vol
       t.float :price_kg
       t.references :carrier, null: false, foreign_key: true
